requirements.txt
streamlit==1.36.0
pandas>=2.0.0
ibm-watsonx-ai
pillow